requirements.txt
langchain==0.2.5
langchain-community==0.2.5
langchain-huggingface==0.0.3
huggingface-hub==0.23.4
faiss-cpu==1.8.0
sentence-transformers==2.7.0
python-dotenv==1.0.1
transformers==4.41.2
langchain_community